requirements.txt
psycopg2-binary~=2.9.9
pandas~=2.2.2
werkzeug~=3.1.3